Rename folder week01 to week04
